refactor(tools): split network device discovery into two tools
Separate discovery from control into distinct tools:

- net_discovery: all scanning/finding mechanisms (DNS, mDNS, ARP, DHCP,
  SSDP, OUI, Zigbee/Z-Wave device listing, unified pipeline)
- net_device_control: hub API interactions for device management (HA
  device listing, entity state reads, service calls)
